updating more terminology
