+++ NModbus/Data/Properties/AssemblyInfo.cs
@@ -9,7 +9,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyCopyright("Copyright © 2006 Scott Alexander, 2015 Dmitry Turin")]
 [assembly: AssemblyDescription("NModbus is a C# implementation of the Modbus protocol. " +
-           "Provides connectivity to Modbus slave compatible devices and applications. " +
+           "Provides connectivity to Modbus server compatible devices and applications. " +
            "Supports ASCII, RTU, TCP, and UDP protocols. ")]
 
 [assembly: AssemblyTrademark("")]
