--- NModbus/Interfaces/ISlaveHandlerContext.cs
@@ -1,7 +1,0 @@
-﻿namespace NModbus
-{
-    public interface ISlaveHandlerContext
-    {
-        IModbusFunctionService GetHandler(byte functionCode);
-    }
-}
