+++ NModbus/Message/WriteFileRecordRequest.cs
@@ -1,13 +1,6 @@
 ﻿using NModbus.Data;
-using NModbus.Unme.Common;
-using NModbus.Utility;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NModbus.Message
 {
@@ -17,8 +10,8 @@ namespace NModbus.Message
         {
         }
 
-        public WriteFileRecordRequest(byte slaveAdress, FileRecordCollection data)
-            : base(slaveAdress, ModbusFunctionCodes.WriteFileRecord)
+        public WriteFileRecordRequest(byte serverAdress, FileRecordCollection data)
+            : base(serverAdress, ModbusFunctionCodes.WriteFileRecord)
         {
             Data = data;
             ByteCount = data.ByteCount;
