+++ README.md
@@ -9,7 +9,7 @@ NModbus
 |**MS .NET**|[![Build status](https://ci.appveyor.com/api/projects/status/a4r06a0owl6xf9ar/branch/NetworkedSlave?svg=true)](https://ci.appveyor.com/project/nmodbus/nmodbus/branch/NetworkedSlave)|
 
 NModbus is a C# implementation of the Modbus protocol.
-Provides connectivity to Modbus slave compatible devices and applications.
+Provides connectivity to Modbus server compatible devices and applications.
 Supports serial ASCII, serial RTU, TCP, and UDP protocols.
 
 History
@@ -22,14 +22,14 @@ The NModbus4 project appears to have gone quiet. This is a fork of that project.
 
 NModbus differs from NModbus4 in following:
 
-- Modbus slave devices are now added to a network which is represented by `IModbusSlaveInstance`.
+- Modbus server devices are now added to a network which is represented by `IModbusServerInstance`.
 - Heavier use of interfaces.
-- Custom function code handlers can be added to slave devices.
+- Custom function code handlers can be added to server devices.
 
 
 Goals
 =======
-- Improve Modbus Slave support (e.g. support multiple slave devices on the same physical transport).
+- Improve Modbus Server support (e.g. support multiple server devices on the same physical transport).
 
 Install
 =======
